dsa: Add support for unmanaged switch mode
Add support for unmanaged switches. Assume that such switches don't
have to be initialized on startup, and that it does not need any tagging
or any other register programming. Provide code for 88e6123_61_65.

This mode is useful if the switch configuration (normally done via eeprom)
does not match the configuration assumptions made in the dsa switch driver.
It is still useful to have such switches supported by DSA to be able to
report switch port statistics.

Unmanaged switch mode is configured in switch platform data.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>